+++ system/libs/pot/OTS_DB_MySQL.php
@@ -98,7 +98,7 @@ class OTS_DB_MySQL extends OTS_Base_DB
 		}
 
 		global $config;
-		$cache = Cache::getInstance();
+		$cache = app()->get('cache');
 		if($cache->enabled()) {
 			$tmp = null;
 			$need_revalidation = true;
@@ -150,7 +150,7 @@ class OTS_DB_MySQL extends OTS_Base_DB
 	{
 		global $config;
 
-		$cache = Cache::getInstance();
+		$cache = app()->get('cache');
 		if($cache->enabled()) {
 			if ($this->clearCacheAfter) {
 				$cache->delete('database_tables');
