+++ system/pages/signature.php
@@ -1,0 +1,6 @@
+<?php
+defined('MYAAC') or die('Direct access not allowed!');
+
+chdir(TOOLS . 'signature');
+include TOOLS . 'signature/index.php';
+exit;
