+++ system/src/App/Admin.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace MyAAC\App;
+
+class Admin
+{
+
+}
