+++ system/src/News.php
@@ -49,7 +49,7 @@ class News
 			'article_image' => ($type == 3 ? $article_image : '')
 		];
 
-		global $hooks;
+		$hooks = app()->get('hooks');
 		if (!$hooks->trigger(HOOK_ADMIN_NEWS_ADD_PRE, $params)) {
 			return false;
 		}
@@ -86,7 +86,7 @@ class News
 			'article_image' => ($type == 3 ? $article_image : ''),
 		];
 
-		global $hooks;
+		$hooks = app()->get('hooks');
 		if (!$hooks->trigger(HOOK_ADMIN_NEWS_UPDATE_PRE, $params)) {
 			return false;
 		}
@@ -105,7 +105,7 @@ class News
 
 	static public function delete($id, &$errors)
 	{
-		global $hooks;
+		$hooks = app()->get('hooks');
 
 		if(isset($id)) {
 			$row = ModelsNews::find($id);
@@ -140,7 +140,7 @@ class News
 
 	static public function toggleHide($id, &$errors, &$status)
 	{
-		global $hooks;
+		$hooks = app()->get('hooks');
 
 		if(isset($id)) {
 			$row = ModelsNews::find($id);
